bindings: network uses entities/types
