+++ pkg/domain/entities/network.go
@@ -3,8 +3,7 @@ package entities
 import (
 	"net"
 
-	commonTypes "github.com/containers/common/libnetwork/types"
-	"github.com/containers/podman/v4/pkg/domain/entities/types"
+	entitiesTypes "github.com/containers/podman/v4/pkg/domain/entities/types"
 )
 
 // NetworkListOptions describes options for listing networks in cli
@@ -22,11 +21,7 @@ type NetworkReloadOptions struct {
 }
 
 // NetworkReloadReport describes the results of reloading a container network.
-type NetworkReloadReport struct {
-	//nolint:stylecheck,revive
-	Id  string
-	Err error
-}
+type NetworkReloadReport = entitiesTypes.NetworkReloadReport
 
 // NetworkRmOptions describes options for removing networks
 type NetworkRmOptions struct {
@@ -35,10 +30,7 @@ type NetworkRmOptions struct {
 }
 
 // NetworkRmReport describes the results of network removal
-type NetworkRmReport struct {
-	Name string
-	Err  error
-}
+type NetworkRmReport = entitiesTypes.NetworkRmReport
 
 // NetworkCreateOptions describes options to create a network
 type NetworkCreateOptions struct {
@@ -68,9 +60,7 @@ type NetworkUpdateOptions struct {
 }
 
 // NetworkCreateReport describes a created network for the cli
-type NetworkCreateReport struct {
-	Name string
-}
+type NetworkCreateReport = entitiesTypes.NetworkCreateReport
 
 // NetworkDisconnectOptions describes options for disconnecting
 // containers from networks
@@ -81,15 +71,12 @@ type NetworkDisconnectOptions struct {
 
 // NetworkConnectOptions describes options for connecting
 // a container to a network
-type NetworkConnectOptions struct {
-	Container string `json:"container"`
-	commonTypes.PerNetworkOptions
-}
+type NetworkConnectOptions = entitiesTypes.NetworkConnectOptions
 
 // NetworkPruneReport containers the name of network and an error
 // associated in its pruning (removal)
 // swagger:model NetworkPruneReport
-type NetworkPruneReport = types.NetworkPruneReport
+type NetworkPruneReport = entitiesTypes.NetworkPruneReport
 
 // NetworkPruneOptions describes options for pruning unused networks
 type NetworkPruneOptions struct {
